Refactor: 라벨 컴포넌트 타입 및 스타일 수정
